Truncate method in calculate class Methods to take User inputs for option 5 & 6 in item menu

+++ src/main/java/bussinessLogic/Item.java
@@ -21,7 +21,7 @@ public class Item {
     public void setItemName(String name){
         this.itemName = name;
     }
-    public void setUnitPrice(float price){
+    public void setUnitPrice(double price){
         this.unitPrice = price;
     }
